Last pre-bisect try: force size_t for CHUNK_SIZE
I strongly doubt an over- or underflow is to blame, but might as well

+++ transformer/src/data/ArchiveParser.hpp
@@ -66,7 +66,7 @@ class Transformer;
 class ArchiveParser {
 private:
     // Approx. 16MB
-    constexpr static auto BLOCK_SIZE = 4194304 * 4;
+    constexpr static auto BLOCK_SIZE = size_t(4194304) * size_t(4);
 
     archive* a;
